requirements.txt
asgiref==3.7.2
attrs==23.1.0
beautifulsoup4==4.12.2
black==23.1.0
boto3==1.26.118
botocore==1.29.165
certifi==2023.7.22
cffi==1.16.0
cfgv==3.4.0
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
cryptography==41.0.4
distlib==0.3.7
dj-rest-auth==3.0.0
Django==4.2.5
django-cors-headers==3.14.0
django-filter==23.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.26.5
exceptiongroup==1.1.3
filelock==3.12.4
flake8==3.9.2
gdown==4.7.1
h11==0.14.0
html2image==2.0.3
identify==2.5.29
idna==3.4
inflection==0.5.1
isort==5.12.0
jmespath==1.0.1
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
mccabe==0.6.1
mypy-extensions==1.0.0
nodeenv==1.8.0
numpy==1.26.0
outcome==1.2.0
packaging==23.1
pandas==2.1.1
pathspec==0.11.2
Pillow==10.0.0
platformdirs==3.10.0
pre-commit==3.4.0
psycopg2-binary==2.9.6
pycodestyle==2.7.0
pycparser==2.21
pydocstyle==6.3.0
pyflakes==2.3.1
PyJWT==2.8.0
pylama==7.7.1
pyOpenSSL==23.2.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.30.2
requests==2.31.0
rpds-py==0.10.3
s3transfer==0.6.2
selenium==4.0.0
six==1.16.0
sniffio==1.3.0
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.4.4
tomli==2.0.1
tqdm==4.66.1
trio==0.22.2
trio-websocket==0.11.1
typing_extensions==4.8.0
tzdata==2023.3
Unidecode==1.3.7
uritemplate==4.1.1
urllib3==1.26.16
urllib3-secure-extra==0.1.0
virtualenv==20.24.5
wsproto==1.2.0
